% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss_case_list.R
\name{pss_case_list}
\alias{pss_case_list}
\title{Create Project SafeSchools Case List from Tracking Log}
\usage{
pss_case_list(tracking_log)
}
\arguments{
\item{tracking_log}{A data frame created from the pss_tracking_log function.}
}
\value{
A data frame for the PSS Case List
}
\description{
Create Project SafeSchools Case List from Tracking Log
}
\examples{
cases <- redcap_read(redcap_uri = url, "token"=token, events = redcap_events)$data
tracking_log <- pss_tracking_log(cases)
case_list <- pss_case_list(tracking_log)
}
