% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss_qc.R
\name{pss_qc}
\alias{pss_qc}
\title{Create QC List}
\usage{
pss_qc(cases, logs, data_dictionary)
}
\arguments{
\item{cases}{A data frame with output from redcap_read function for PSS.}

\item{logs}{A data frame with all logging events from REDCap.}

\item{data_dictionary}{A data frame with the current data dictionary to look for missingness.}
}
\value{
A cleaned data frame of current QC Issues.
}
\description{
Create QC List
}
\examples{
cases <- redcap_read(redcap_uri = url, "token"=token, events = redcap_events)$data
qc <- pss_qc(cases)
}
