% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pss_tacking_log.R
\name{pss_tracking_log}
\alias{pss_tracking_log}
\title{Create PSS Tracking Log From REDCap Data}
\usage{
pss_tracking_log(cases, week_start_day = 5)
}
\arguments{
\item{cases}{A data frame with output from redcap_read function for PSS}
}
\value{
A cleaned tracking log data frame.
}
\description{
Create PSS Tracking Log From REDCap Data
}
\examples{
cases <- redcap_read(redcap_uri = url, "token"=token, events = redcap_events)$data
tracking_log <- pss_tracking_log(cases)

}
