% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_one}
\alias{str_split_one}
\title{String Split 1}
\usage{
str_split_one(x, split)
}
\arguments{
\item{x}{A character vector with one element.}

\item{split}{What to split on.}
}
\value{
A character vector.
}
\description{
String Split 1
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit(x, split = ",")
}
